Memberlist Changes
Memberlist_Head:
Mostly done, 3 lines untranslated since I don't know where thay are in-game to see what they might be, "Limit" has been translated but not seen it in-game so might not look right.

Memberlist_Tail:
Translated the elements that are also found on _Head but not sure where they show up.